fix: parameters on Whisper demo class and renaming of module
